retrieval-augmented-generation/requirements.txt
openai
huggingface-hub
numpy
pandas
tqdm
datasets
python-dotenv==1.0.1


